ICBM 0.5.8 Release

+++ src/minecraft/icbm/SmokeFX.java
@@ -6,7 +6,7 @@ import universalelectricity.prefab.Vector3;
 
 public class SmokeFX extends EntitySmokeFX
 {
-	public SmokeFX(World par1World, Vector3 position, float red, float green, float blue, float scale, float distance)
+	public SmokeFX(World par1World, Vector3 position, float red, float green, float blue, float scale, double distance)
 	{
 		super(par1World, position.x, position.y, position.z, 0, 0, 0, scale);
 		this.renderDistanceWeight = distance;
